requirements.txt
spotipy
python-dotenv
tqdm
py-sonic
musicbrainzngs
thefuzz
pytest
build
twine
streamrip
ytmusicapi
click

fastapi[standard]
sqlmodel
pydantic-settings
bcrypt
pyjwt
aiomysql
greenlet
httpx
cryptography